replace old process module with the new one
