+++ eo_tools/S1/process.py
@@ -50,6 +50,7 @@ def process_insar(
     filter_ifg: bool = True,
     multilook: List[int] = [1, 4],
     warp_kernel: str = "bicubic",
+    cal_type: str = "beta",
     clip_to_shape: bool = True,
     skip_preprocessing: bool = False,
 ) -> str:
@@ -76,6 +77,7 @@ def process_insar(
         filter_ifg (bool): Also applies boxcar to interferogram. Has no effect if file_complex_ifg is set to None or write_coherence is set to False. Defaults to True.x
         multilook (List[int], optional): Multilooking to apply prior to geocoding. Defaults to [1, 4].
         warp_kernel (str, optional): Resampling kernel used in coregistration and geocoding. Possible values are "nearest", "bilinear", "bicubic" and "bicubic6". Defaults to "bicubic".
+        cal_type (str, optional): Type of radiometric calibration. "beta" or "sigma" nought. Defaults to "beta" 
         clip_to_shape (bool, optional): If set to False the geocoded images are not clipped according to the `shp` parameter. They are made of all the bursts intersecting the `shp` geometry. Defaults to True.
         skip_preprocessing (bool, optional): Skip the processing part in case the files are already written. Defaults to False.
 
@@ -98,6 +100,8 @@ def process_insar(
         pol=pol,
         apply_fast_esd=apply_fast_esd,
         subswaths=subswaths,
+        warp_kernel=warp_kernel,
+        cal_type=cal_type,
         dir_dem=dir_dem,
         dem_upsampling=dem_upsampling,
         dem_force_download=dem_force_download,
@@ -114,8 +118,6 @@ def process_insar(
         var_names.append("amp_prm")
     if write_secondary_amplitude:
         var_names.append("amp_sec")
-
-
 
     if isinstance(pol, str):
         if pol == "full":
@@ -182,6 +184,105 @@ def process_insar(
     return Path(out_dir).parent
 
 
+def process_slc(
+    dir_prm: str,
+    outputs_prefix: str,
+    aoi_name: str = None,
+    shp: shape = None,
+    pol: Union[str, List[str]] = "full",
+    subswaths: List[str] = ["IW1", "IW2", "IW3"],
+    dir_dem: str = "/tmp",
+    dem_upsampling: float = 1.8,
+    dem_force_download: bool = True,
+    dem_buffer_arc_sec: float = 40,
+    multilook: List[int] = [1, 4],
+    warp_kernel: str = "bicubic",
+    cal_type: str = "sigmaNought",
+    clip_to_shape: bool = True,
+    skip_preprocessing: bool = False,
+) -> str:
+    """Performs InSAR processing of a pair of SLC Sentinel-1 products, geocode the outputs and writes them as COG (Cloud Optimized GeoTiFF) files.
+    AOI crop is optional.
+
+    Args:
+        dir_prm (str): primary image (SLC Sentinel-1 product directory).
+        outputs_prefix (str): location in which the product subdirectory will be created
+        aoi_name (str, optional): optional suffix to describe AOI / experiment. Defaults to None.
+        shp (shapely.geometry.shape, optional): Shapely geometry describing an area of interest as a polygon. Defaults to None.
+        pol (Union[str, List[str]], optional): Polarimetric channels to process (Either 'VH','VV, 'full' or a list like ['HV', 'VV']).  Defaults to "full".
+        subswaths (List[str], optional): limit the processing to a list of subswaths like `["IW1", "IW2"]`. Defaults to ["IW1", "IW2", "IW3"].
+        dem_upsampling (float, optional): upsampling factor for the DEM, it is recommended to keep the default value. Defaults to 1.8.
+        dem_force_download (bool, optional):  To reduce execution time, DEM files are stored on disk. Set to True to redownload these files if necessary. Defaults to False.
+        dem_buffer_arc_sec (float, optional): Increase if the image area is not completely inside the DEM. Defaults to 40.
+        multilook (List[int], optional): Multilooking to apply prior to geocoding. Defaults to [1, 4].
+        warp_kernel (str, optional): Resampling kernel used in coregistration and geocoding. Possible values are "nearest", "bilinear", "bicubic" and "bicubic6". Defaults to "bicubic".
+        cal_type (str, optional): Type of radiometric calibration. "beta" or "sigma" nought. Defaults to "beta" 
+        clip_to_shape (bool, optional): If set to False the geocoded images are not clipped according to the `shp` parameter. They are made of all the bursts intersecting the `shp` geometry. Defaults to True.
+        skip_preprocessing (bool, optional): Skip the processing part in case the files are already written. Defaults to False.
+
+    Returns:
+        str: output directory
+    """
+
+    if not np.any([coherence, interferogram]):
+        raise ValueError(
+            "At least one of `write_coherence` and `write_interferogram` must be True."
+        )
+
+    # prepare pair for interferogram computation
+    out_dir = prepare_slc(
+        dir_prm=dir_prm,
+        outputs_prefix=outputs_prefix,
+        aoi_name=aoi_name,
+        shp=shp,
+        pol=pol,
+        subswaths=subswaths,
+        cal_type=cal_type,
+        dir_dem=dir_dem,
+        dem_upsampling=dem_upsampling,
+        dem_force_download=dem_force_download,
+        dem_buffer_arc_sec=dem_buffer_arc_sec,
+        skip_preprocessing=skip_preprocessing,
+    )
+
+    var_names = []
+    var_names.append("amp_prm")
+
+    if isinstance(pol, str):
+        if pol == "full":
+            pol_ = ["vv", "vh"]
+        else:
+            pol_ = [pol.lower()]
+    elif isinstance(pol, list):
+        pol_ = [p.lower() for p in pol]
+
+    iw_idx = [iw[2] for iw in subswaths]
+    patterns = [f"{p}_iw{iw}" for p in pol_ for iw in iw_idx]
+    for pattern in patterns:
+        file_prm = f"{out_dir}/slc_prm_{pattern}.tif"
+
+        if os.path.isfile(file_prm):
+            log.info(
+                f"---- Amplitude for {" ".join(pattern.split('/')[-1].split('_')).upper()}"
+            )
+
+            file_ampl = f"{out_dir}/amp_prm_{pattern}.tif"
+            amplitude(file_in=file_prm, file_out=file_ampl)
+
+    # by default, we use iw and pol which exist
+    geocode_and_merge_iw(
+        input_dir=Path(out_dir).parent,
+        var_names=var_names,
+        shp=shp,
+        pol=["vv", "vh"],
+        subswaths=["IW1", "IW2", "IW3"],
+        multilook=multilook,
+        warp_kernel=warp_kernel,
+        clip_to_shape=clip_to_shape,
+    )
+    return Path(out_dir).parent
+
+
 def prepare_insar(
     dir_prm: str,
     dir_sec: str,
@@ -192,6 +293,7 @@ def prepare_insar(
     subswaths: List[str] = ["IW1", "IW2", "IW3"],
     apply_fast_esd: bool = False,
     warp_kernel: str = "bicubic",
+    cal_type: str = "beta",
     dir_dem: str = "/tmp",
     dem_upsampling: float = 1.8,
     dem_force_download: bool = True,
@@ -209,6 +311,8 @@ def prepare_insar(
         pol (Union[str, List[str]], optional):  Polarimetric channels to process (Either 'VH','VV, 'full' or a list like ['HV', 'VV']).  Defaults to "full".
         subswaths (List[str], optional):  limit the processing to a list of subswaths like `["IW1", "IW2"]`. Defaults to ["IW1", "IW2", "IW3"].
         apply_fast_esd (bool, optional): correct the phase to avoid jumps between bursts. This has no effect if only one burst is processed.  Defaults to False.
+        warp_kernel (str, optional): kernel used to align secondary SLC. Possible values are "nearest", "bilinear", "bicubic" and "bicubic6".Defaults to "bilinear".
+        cal_type (str, optional): Type of radiometric calibration. "beta" or "sigma" nought. Defaults to "beta" 
         dem_upsampling (float, optional): upsampling factor for the DEM, it is recommended to keep the default value. Defaults to 1.8.
         dem_force_download (bool, optional):   To reduce execution time, DEM files are stored on disk. Set to True to redownload these files if necessary. Defaults to True.
         dem_buffer_arc_sec (float, optional): Increase if the image area is not completely inside the DEM. Defaults to 40.
@@ -313,6 +417,7 @@ def prepare_insar(
                     max_burst=burst_prm_max,
                     apply_fast_esd=apply_fast_esd,
                     warp_kernel=warp_kernel,
+                    cal_type=cal_type,
                     dir_dem=dir_dem,
                     dem_upsampling=dem_upsampling,
                     dem_buffer_arc_sec=dem_buffer_arc_sec,
@@ -325,6 +430,133 @@ def prepare_insar(
                 os.rename(
                     f"{out_dir}/secondary.tif",
                     f"{out_dir}/slc_sec_{p.lower()}_iw{iw}.tif",
+                )
+                os.rename(f"{out_dir}/lut.tif", f"{out_dir}/lut_{p.lower()}_iw{iw}.tif")
+            else:
+                log.info("Skipping preprocessing.")
+    return out_dir
+
+
+def prepare_slc(
+    dir_prm: str,
+    outputs_prefix: str,
+    aoi_name: str = None,
+    shp: shape = None,
+    pol: Union[str, List[str]] = "full",
+    subswaths: List[str] = ["IW1", "IW2", "IW3"],
+    cal_type: str = "beta",
+    # warp_kernel: str = "bicubic",
+    dir_dem: str = "/tmp",
+    dem_upsampling: float = 1.8,
+    dem_force_download: bool = True,
+    dem_buffer_arc_sec: float = 40,
+    skip_preprocessing: bool = False,
+) -> str:
+    """Produce a coregistered pair of Single Look Complex images and associated lookup tables.
+
+    Args:
+        dir_prm (str): Primary image (SLC Sentinel-1 product directory).
+        outputs_prefix (str): location in which the product subdirectory will be created.
+        aoi_name (str, optional): optional suffix to describe AOI / experiment. Defaults to None.
+        shp (shapely.geometry.shape, optional): Shapely geometry describing an area of interest as a polygon. Defaults to None.
+        pol (Union[str, List[str]], optional):  Polarimetric channels to process (Either 'VH','VV, 'full' or a list like ['HV', 'VV']).  Defaults to "full".
+        subswaths (List[str], optional):  limit the processing to a list of subswaths like `["IW1", "IW2"]`. Defaults to ["IW1", "IW2", "IW3"].
+        dem_upsampling (float, optional): upsampling factor for the DEM, it is recommended to keep the default value. Defaults to 1.8.
+        dem_force_download (bool, optional):   To reduce execution time, DEM files are stored on disk. Set to True to redownload these files if necessary. Defaults to True.
+        dem_buffer_arc_sec (float, optional): Increase if the image area is not completely inside the DEM. Defaults to 40.
+        skip_preprocessing (bool, optional): Skip the processing part in case the files are already written. It is recommended to leave this parameter to default value. Defaults to False.
+
+    Returns:
+        str: output directory
+    """
+
+    if aoi_name is None:
+        aoi_substr = ""
+    else:
+        aoi_substr = f"_{aoi_name}"
+
+    if not isinstance(subswaths, list):
+        raise ValueError("Subswaths must be a list like ['IW1', 'IW2'].")
+
+    # retrieve burst geometries
+    gdf_burst_prm = get_burst_geometry(
+        dir_prm, target_subswaths=["IW1", "IW2", "IW3"], polarization="VV"
+    )
+
+    # find what subswaths and bursts intersect AOI
+    if shp is not None:
+        gdf_burst_prm = gdf_burst_prm[gdf_burst_prm.intersects(shp)]
+
+    if gdf_burst_prm.empty:
+        raise RuntimeError(
+            "The list of bursts to process is empty. Make sure shp intersects the product."
+        )
+
+    # identify corresponding subswaths
+    sel_subsw_prm = gdf_burst_prm["subswath"]
+    unique_subswaths = np.unique(sel_subsw_prm)
+    unique_subswaths = [it for it in unique_subswaths if it in subswaths]
+
+    # check that polarization is correct
+    info_prm = identify(dir_prm)
+    if isinstance(pol, str):
+        if pol == "full":
+            pol_ = info_prm.polarizations
+        else:
+            if pol.upper() in info_prm.polarizations:
+                pol_ = [pol]
+            else:
+                raise RuntimeError(
+                    f"polarization {pol} does not exists in the source product"
+                )
+    elif isinstance(pol, list):
+        pol_ = [x for x in pol if x in info_prm.polarizations]
+    else:
+        raise RuntimeError("polarizations must be of type str or list")
+
+    meta_prm = info_prm.scanMetadata()
+    # parse dates
+    datestr_prm = meta_prm["start"]
+    date_prm = datetime.strptime(datestr_prm, "%Y%m%dT%H%M%S")
+
+    id_prm = date_prm.strftime("%Y-%m-%d-%H%M%S")
+
+    out_dir = f"{outputs_prefix}/S1_{id_prm}{aoi_substr}/sar"
+    if not os.path.isdir(out_dir):
+        log.info(f"Creating directory {out_dir}")
+        os.makedirs(out_dir)
+    for p in pol_:
+        for subswath in unique_subswaths:
+            log.info(f"---- Processing subswath {subswath} in {p.upper()} polarization")
+
+            # identify bursts to process
+            bursts_prm = gdf_burst_prm[gdf_burst_prm["subswath"] == subswath][
+                "burst"
+            ].values
+            burst_prm_min = bursts_prm.min()
+            burst_prm_max = bursts_prm.max()
+
+            iw = int(subswath[2])
+            if not os.path.isdir(out_dir):
+                os.mkdir(out_dir)
+            if not skip_preprocessing:
+                preprocess_slc_iw(
+                    dir_prm,
+                    out_dir,
+                    iw=iw,
+                    pol=p.lower(),
+                    min_burst=burst_prm_min,
+                    max_burst=burst_prm_max,
+                    cal_type=cal_type,
+                    # warp_kernel=warp_kernel,
+                    dir_dem=dir_dem,
+                    dem_upsampling=dem_upsampling,
+                    dem_buffer_arc_sec=dem_buffer_arc_sec,
+                    dem_force_download=dem_force_download,
+                )
+                os.rename(
+                    f"{out_dir}/primary.tif",
+                    f"{out_dir}/slc_prm_{p.lower()}_iw{iw}.tif",
                 )
                 os.rename(f"{out_dir}/lut.tif", f"{out_dir}/lut_{p.lower()}_iw{iw}.tif")
             else:
@@ -464,6 +696,7 @@ def preprocess_insar_iw(
     max_burst: int = None,
     apply_fast_esd: bool = True,
     warp_kernel: str = "bicubic",
+    cal_type: str = "beta",
     dir_dem: str = "/tmp",
     dem_upsampling: float = 1.8,
     dem_buffer_arc_sec: float = 40,
@@ -483,6 +716,7 @@ def preprocess_insar_iw(
         dir_dem (str, optional): directory where the DEM is downloaded. Must be created beforehand. Defaults to "/tmp".
         apply_fast_esd: (bool, optional): correct the phase to avoid jumps between bursts. This has no effect if only one burst is processed. Defaults to True.
         warp_kernel (str, optional): kernel used to align secondary SLC. Possible values are "nearest", "bilinear", "bicubic" and "bicubic6".Defaults to "bilinear".
+        cal_type (str, optional): Type of radiometric calibration. "beta" or "sigma" nought. Defaults to "beta" 
         dem_upsampling (float, optional): Upsample the DEM, it is recommended to keep the default value. Defaults to 2.
         dem_buffer_arc_sec (float, optional): Increase if the image area is not completely inside the DEM. Defaults to 40.
         dem_force_download (bool, optional): To reduce execution time, DEM files are stored on disk. Set to True to redownload these files if necessary. Defaults to false.
@@ -499,9 +733,11 @@ def preprocess_insar_iw(
 
     if pol not in ["vv", "vh"]:
         ValueError("pol must be 'vv' or 'vh'")
-    
-    if dem_force_download:
-        log.warning("dem_force_download is disabled. This could result in wrong outputs if the file on disk does not match dem_upsampling and dem_buffer_arc_sec.")
+
+    if not dem_force_download:
+        log.warning(
+            "dem_force_download is disabled. This could result in wrong outputs if the file on disk does not match dem_upsampling and dem_buffer_arc_sec."
+        )
 
     prm = S1IWSwath(dir_primary, iw=iw, pol=pol)
     sec = S1IWSwath(dir_secondary, iw=iw, pol=pol)
@@ -554,7 +790,7 @@ def preprocess_insar_iw(
 
     warnings.filterwarnings("ignore", category=rio.errors.NotGeoreferencedWarning)
     _child_process(
-        _process_bursts,
+        _process_bursts_insar,
         (
             prm,
             sec,
@@ -571,6 +807,7 @@ def preprocess_insar_iw(
             dem_force_download,
             warp_kernel,
             overlap,
+            cal_type,
         ),
     )
 
@@ -617,6 +854,126 @@ def preprocess_insar_iw(
             os.remove(tmp_prm)
         if os.path.isfile(tmp_sec):
             os.remove(tmp_sec)
+
+    log.info("Done")
+
+
+def preprocess_slc_iw(
+    dir_primary: str,
+    dir_out: str,
+    iw: int = 1,
+    pol: Union[str, List[str]] = "vv",
+    min_burst: int = 1,
+    max_burst: int = None,
+    cal_type: str = "beta",
+    dir_dem: str = "/tmp",
+    dem_upsampling: float = 1.8,
+    dem_buffer_arc_sec: float = 40,
+    dem_force_download: bool = True,
+) -> None:
+    """Pre-process S1 InSAR subswaths pairs. Write coregistered primary and secondary SLC files as well as a lookup table that can be used to geocode rasters in the single-look radar geometry.
+
+    Args:
+        dir_primary (str): directory containing the primary SLC product of the pair.
+        dir_out (str): output directory (creating it if does not exist).
+        dir_dem (str, optional): directory where DEMs used for geocoding are stored. Defaults to "/tmp".
+        iw (int, optional): subswath index. Defaults to 1.
+        pol (str, optional): polarization ('vv','vh'). Defaults to "vv".
+        min_burst (int, optional): first burst to process. Defaults to 1.
+        max_burst (int, optional): fast burst to process. If not set, last burst of the subswath. Defaults to None.
+        dir_dem (str, optional): directory where the DEM is downloaded. Must be created beforehand. Defaults to "/tmp".
+        warp_kernel (str, optional): kernel used to align secondary SLC. Possible values are "nearest", "bilinear", "bicubic" and "bicubic6".Defaults to "bilinear".
+        dem_upsampling (float, optional): Upsample the DEM, it is recommended to keep the default value. Defaults to 2.
+        dem_buffer_arc_sec (float, optional): Increase if the image area is not completely inside the DEM. Defaults to 40.
+        dem_force_download (bool, optional): To reduce execution time, DEM files are stored on disk. Set to True to redownload these files if necessary. Defaults to false.
+
+    Note:
+        DEM-assisted coregistration is performed to align the secondary with the Primary. A lookup table file is written to allow the geocoding images from the radar (single-look) grid to the geographic coordinates of the DEM. Bursts are stitched together to form continuous images. All output files are in the GeoTiff format that can be handled by most GIS softwares and geospatial raster tools such as GDAL and rasterio. Because they are in the SAR geometry, SLC rasters are not georeferenced.
+    """
+
+    if not os.path.isdir(dir_out):
+        os.mkdir(dir_out)
+
+    if iw not in [1, 2, 3]:
+        ValueError("iw must be 1, 2 or 3")
+
+    if pol not in ["vv", "vh"]:
+        ValueError("pol must be 'vv' or 'vh'")
+
+    if not dem_force_download:
+        log.warning(
+            "dem_force_download is disabled. This could result in wrong outputs if the file on disk does not match dem_upsampling and dem_buffer_arc_sec."
+        )
+
+    prm = S1IWSwath(dir_primary, iw=iw, pol=pol)
+    # sec = S1IWSwath(dir_secondary, iw=iw, pol=pol)
+
+    prm_burst_info = prm.meta["product"]["swathTiming"]["burstList"]["burst"]
+    # sec_burst_info = sec.meta["product"]["swathTiming"]["burstList"]["burst"]
+
+    overlap = np.round(prm.compute_burst_overlap(2)).astype(int)
+
+    if not max_burst:
+        max_burst_ = prm.burst_count
+    else:
+        max_burst_ = max_burst
+
+    if max_burst_ > min_burst:
+        tmp_prm = f"{dir_out}/tmp_primary.tif"
+    elif max_burst_ < min_burst:
+        raise ValueError("max_burst must be >= min_burst")
+    else:
+        tmp_prm = f"{dir_out}/primary.tif"
+
+    if (
+        max_burst_ > prm.burst_count
+        or max_burst_ < 1
+        or min_burst > prm.burst_count
+        or min_burst < 1
+    ):
+        raise ValueError(
+            f"min_burst and max_burst must be values between 1 and {prm.burst_count}"
+        )
+
+    naz = prm.lines_per_burst * (max_burst_ - min_burst + 1)
+    nrg = prm.samples_per_burst
+
+    warnings.filterwarnings("ignore", category=rio.errors.NotGeoreferencedWarning)
+    _child_process(
+        _process_bursts_slc,
+        (
+            prm,
+            tmp_prm,
+            dir_out,
+            dir_dem,
+            naz,
+            nrg,
+            min_burst,
+            max_burst_,
+            dem_upsampling,
+            dem_buffer_arc_sec,
+            dem_force_download,
+            overlap,
+            cal_type
+        ),
+    )
+
+    if max_burst_ > min_burst:
+        _child_process(
+            _stitch_bursts,
+            (
+                tmp_prm,
+                f"{dir_out}/primary.tif",
+                prm.lines_per_burst,
+                max_burst_ - min_burst + 1,
+                overlap,
+            ),
+        )
+
+    log.info("Cleaning temporary files")
+    if max_burst_ > min_burst:
+        if os.path.isfile(tmp_prm):
+            os.remove(tmp_prm)
 
     log.info("Done")
 
@@ -864,7 +1221,7 @@ def coherence(
 # Auxiliary functions which are not supposed to be used outside of the processor
 
 
-def _process_bursts(
+def _process_bursts_insar(
     prm,
     sec,
     tmp_prm,
@@ -880,6 +1237,7 @@ def _process_bursts(
     dem_force_download,
     warp_kernel,
     overlap,
+    cal_type
 ):
 
     H = int(overlap / 2)
@@ -974,6 +1332,12 @@ def _process_bursts(
             arr_p = prm.read_burst(burst_idx, True)
             arr_s = sec.read_burst(burst_idx, True)
 
+            # calibration (beta or sigma nought)
+            cal_p = prm.calibration_factor(burst_idx, cal_type=cal_type)
+            arr_p /= cal_p
+            cal_s = sec.calibration_factor(burst_idx, cal_type=cal_type)
+            arr_s /= cal_s
+
             # deramp secondary
             pdb_s = sec.deramp_burst(burst_idx)
             arr_s *= np.exp(1j * pdb_s)
@@ -1022,7 +1386,131 @@ def _process_bursts(
 
     with rio.open(file_lut, "w", **prof_lut) as ds_lut:
         ds_lut.write(arr_lut)
-    
+
+
+def _process_bursts_slc(
+    prm,
+    tmp_prm,
+    dir_out,
+    dir_dem,
+    naz,
+    nrg,
+    min_burst,
+    max_burst,
+    dem_upsampling,
+    dem_buffer_arc_sec,
+    dem_force_download,
+    overlap,
+    cal_type,
+):
+
+    H = int(overlap / 2)
+    prof_tmp = dict(
+        width=nrg,
+        height=naz,
+        count=1,
+        dtype="complex64",
+        driver="GTiff",
+        nodata=np.nan,
+        # compress="zstd",
+        # num_threads="all_cpus",
+        tiled=True,
+        blockxsize=512,
+        blockysize=512,
+    )
+    warnings.filterwarnings("ignore", category=rio.errors.NotGeoreferencedWarning)
+    # process individual bursts
+    file_dem = prm.fetch_dem(
+        min_burst,
+        max_burst,
+        dir_dem,
+        buffer_arc_sec=dem_buffer_arc_sec,
+        force_download=dem_force_download,
+        upscale_factor=dem_upsampling,
+    )
+    file_lut = f"{dir_out}/lut.tif"
+    with rio.open(file_dem) as ds_dem:
+        width_lut = ds_dem.width
+        height_lut = ds_dem.height
+        crs_lut = ds_dem.crs
+        transform_lut = ds_dem.transform
+
+    prof_lut = dict(
+        width=width_lut,
+        height=height_lut,
+        count=2,
+        dtype=np.float64,
+        crs=crs_lut,
+        transform=transform_lut,
+        nodata=np.nan,
+        tiled=True,
+        blockxsize=512,
+        blockysize=512,
+    )
+
+    arr_lut = np.full((2, height_lut, width_lut), fill_value=np.nan)
+
+    with rio.open(tmp_prm, "w", **prof_tmp) as ds_prm:
+        off_az = 0
+        for burst_idx in range(min_burst, max_burst + 1):
+            log.info(f"---- Processing burst {burst_idx} ----")
+
+            # compute geocoding LUTs (lookup tables) for primary and secondary bursts
+            file_dem_burst = f"{dir_out}/dem_burst.tif"
+            burst_geoms = prm.gdf_burst_geom
+            burst_geom = burst_geoms[burst_geoms["burst"] == burst_idx].iloc[0]
+            shp = burst_geom.geometry.buffer(dem_buffer_arc_sec / 3600)
+
+            with rio.open(file_dem) as ds_dem:
+                w = geometry_window(ds_dem, shapes=[shp])
+                # window to read in the DEM
+                burst_window = [w.col_off, w.row_off, w.width, w.height]
+                # pixel position to write burst in the LUT
+                slices = w.toslices()
+
+            # use virtual raster to keep using the same geocoding function
+            file_dem_burst = f"{dir_out}/dem_burst.vrt"
+            gdal.Translate(
+                destName=file_dem_burst,
+                srcDS=file_dem,
+                format="VRT",
+                srcWin=burst_window,
+                creationOptions=["BLOCKXSIZE=512", "BLOCKYSIZE=512"],
+            )
+
+            # this implementation upsamples DEM at download, not during geocoding
+            az_p2g, rg_p2g, _ = prm.geocode_burst(
+                file_dem_burst,
+                burst_idx=burst_idx,
+                dem_upsampling=1,
+            )
+
+            # read primary and secondary burst rasters
+            arr_p = prm.read_burst(burst_idx, True)
+            cal_p = prm.calibration_factor(burst_idx, cal_type=cal_type)
+            arr_p /= cal_p
+
+            first_line = (burst_idx - min_burst) * prm.lines_per_burst
+
+            # write the coregistered SLCs
+            ds_prm.write(
+                arr_p, 1, window=Window(0, first_line, nrg, prm.lines_per_burst)
+            )
+
+            # place overlapping burst LUT with azimuth offset
+            if burst_idx > min_burst:
+                msk_overlap = az_p2g < H
+                az_p2g[msk_overlap] = np.nan
+                rg_p2g[msk_overlap] = np.nan
+            msk = ~np.isnan(az_p2g)
+            arr_lut[0, slices[0], slices[1]][msk] = az_p2g[msk] + off_az
+            arr_lut[1, slices[0], slices[1]][msk] = rg_p2g[msk]
+            off_az += prm.lines_per_burst - 2 * H
+
+    remove(file_dem_burst)
+
+    with rio.open(file_lut, "w", **prof_lut) as ds_lut:
+        ds_lut.write(arr_lut)
 
 
 def _apply_fast_esd(
